+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ChangeLog history
 ## Deprecations 
 
 ## Features
+- quit neuroglancer by enter q and return.
 - work with Flatiron disBatch.
 - a new operator remap to renumber a serials of segmentation chunks.
 - shard meshing. It is not validated yet. The data is written, but Neuroglancer is still not showing them. There is something like manifest missing.